refactor: cleanup feature dev imports in transform (#5395)
